move data sets to a hidden file
The hidden file, called .hidden, will be ignored by Git and will not be used when building the package. Therefore, it only exists in my local repo.

As a result of this change, some of the tests and files will not be run properly because the data sets do not exist.