fix(get_OTcost): remove use of reference matrix for cost
The `transport` package previously reported the `from` and `to` indices
of the transport map according to the vectors of positive entries in the
pre- and post-distributions.
On August 10, 2022, a bug fix changed this so that the indices referred
to the entire pre- and post-vectors.

See https://cran.r-project.org/web/packages/transport/NEWS for details
on the bug report.